requirements.txt
livekit-agents[openai, tavus, deepgram, rime]==1.0.17
python-dotenv~=1.0